+++ pkg/exporter/refs.go
@@ -82,6 +82,7 @@ func pullRefsFromProject(p schemas.Project) error {
 			strings.Join(gp.TagList, ","),
 			p.OutputSparseStatusMetrics(),
 			p.Pull.Pipeline.Jobs.Enabled(),
+			p.Pull.Pipeline.Jobs.FromChildPipelines.Enabled(),
 			p.Pull.Pipeline.Variables.Enabled(),
 			p.Pull.Pipeline.Variables.Regexp(),
 		)